// src/app/videojuego/interfaces/informacion_estudiante.interface.ts
export default interface Evaluacion_Estudiante {
  id_estudiante: number;
  eficiencia_algoritmica: number;
  reconocimiento_patrones: string;
  identificacion_errores: string;
  abstraccion: string;
  asociacion: string;
  construccion_algoritmos: string;
  p_actividades_completadas: number;
  tipo_premiacion: string; // o string[], si es un arreglo
}